feat: Modularize Clean Architecture into core, feature, app (v2)
Re-applies the modularization of the Clean Architecture implementation into distinct Gradle modules, based on the main branch:

- `core`: Contains shared components: MVI core, base DI modules (Network, Database), common data models, and database definitions.
- `feature_welcome`: Encapsulates domain, data, and presentation layers for the "welcome message" functionality, including its DI module.
- `app`: Main application module, responsible for application entry points and global configurations. Depends on core and feature modules. Redundant dependencies (Room, Moshi) were commented out from its Gradle file.

This refactoring enhances code organization, separation of concerns, and potential build time improvements.

Note: I was unable to verify the build or run tests in this environment.
