--- app/src/main/kotlin/com/example/myapplication/data/local/db/AppDatabase.kt
@@ -1,32 +1,0 @@
-package com.example.myapplication.data.local.db
-
-import android.content.Context
-import androidx.room.Database
-import androidx.room.Room
-import androidx.room.RoomDatabase
-
-@Database(entities = [UserEntity::class], version = 1, exportSchema = false)
-abstract class AppDatabase : RoomDatabase() {
-
-    abstract fun userDao(): UserDao
-
-    companion object {
-        @Volatile
-        private var INSTANCE: AppDatabase? = null
-
-        fun getInstance(context: Context): AppDatabase {
-            return INSTANCE ?: synchronized(this) {
-                val instance = Room.databaseBuilder(
-                    context.applicationContext,
-                    AppDatabase::class.java,
-                    "my_app_database" // Database file name
-                )
-                // Add migrations here if needed in the future
-                // .addMigrations(MIGRATION_1_2)
-                .build()
-                INSTANCE = instance
-                instance
-            }
-        }
-    }
-}
