--- app/src/main/kotlin/com/example/myapplication/data/local/db/UserDao.kt
@@ -1,22 +1,0 @@
-package com.example.myapplication.data.local.db
-
-import androidx.room.Dao
-import androidx.room.Insert
-import androidx.room.OnConflictStrategy
-import androidx.room.Query
-import kotlinx.coroutines.flow.Flow
-
-@Dao
-interface UserDao {
-    @Query("SELECT * FROM users")
-    fun getAllUsers(): Flow<List<UserEntity>>
-
-    @Query("SELECT * FROM users WHERE id = :userId")
-    suspend fun getUserById(userId: Int): UserEntity?
-
-    @Insert(onConflict = OnConflictStrategy.REPLACE)
-    suspend fun insertUser(user: UserEntity)
-
-    @Query("DELETE FROM users WHERE id = :userId")
-    suspend fun deleteUserById(userId: Int)
-}
