--- app/src/main/kotlin/com/example/myapplication/data/local/db/UserEntity.kt
@@ -1,12 +1,0 @@
-package com.example.myapplication.data.local.db
-
-import androidx.room.Entity
-import androidx.room.PrimaryKey
-
-@Entity(tableName = "users")
-data class UserEntity(
-    @PrimaryKey(autoGenerate = true)
-    val id: Int = 0,
-    val name: String,
-    val email: String? = null
-)
