--- app/src/main/kotlin/com/example/myapplication/data/remote/model/UserDto.kt
@@ -1,16 +1,0 @@
-package com.example.myapplication.data.remote.model
-
-import com.squareup.moshi.Json
-import com.squareup.moshi.JsonClass
-
-@JsonClass(generateAdapter = true)
-data class UserDto(
-    @Json(name = "id")
-    val id: Long,
-    @Json(name = "name")
-    val name: String,
-    @Json(name = "username")
-    val username: String,
-    @Json(name = "email")
-    val email: String
-)
