--- app/src/main/kotlin/com/example/myapplication/di/DatabaseModule.kt
@@ -1,28 +1,0 @@
-package com.example.myapplication.di
-
-import android.content.Context
-import com.example.myapplication.data.local.db.AppDatabase
-import com.example.myapplication.data.local.db.UserDao
-import dagger.Module
-import dagger.Provides
-import dagger.hilt.InstallIn
-import dagger.hilt.android.qualifiers.ApplicationContext
-import dagger.hilt.components.SingletonComponent
-import javax.inject.Singleton
-
-@Module
-@InstallIn(SingletonComponent::class)
-object DatabaseModule {
-
-    @Provides
-    @Singleton
-    fun provideAppDatabase(@ApplicationContext appContext: Context): AppDatabase {
-        return AppDatabase.getInstance(appContext)
-    }
-
-    @Provides
-    @Singleton // Daos are usually singletons tied to the database instance
-    fun provideUserDao(appDatabase: AppDatabase): UserDao {
-        return appDatabase.userDao()
-    }
-}
