--- app/src/main/kotlin/com/example/myapplication/di/NetworkModule.kt
@@ -1,24 +1,0 @@
-package com.example.myapplication.di
-
-import com.squareup.moshi.Moshi
-import com.squareup.moshi.kotlin.reflect.KotlinJsonAdapterFactory
-import dagger.Module
-import dagger.Provides
-import dagger.hilt.InstallIn
-import dagger.hilt.components.SingletonComponent
-import javax.inject.Singleton
-
-@Module
-@InstallIn(SingletonComponent::class)
-object NetworkModule {
-
-    @Provides
-    @Singleton
-    fun provideMoshi(): Moshi {
-        return Moshi.Builder()
-            .add(KotlinJsonAdapterFactory()) // For Kotlin support
-            .build()
-    }
-
-    // You would add other network-related providers here, like Retrofit, OkHttpClient, etc.
-}
