--- app/src/main/kotlin/com/example/myapplication/domain/model/Message.kt
@@ -1,5 +1,0 @@
-package com.example.myapplication.domain.model
-
-data class Message(
-    val content: String
-)
