+++ core/consumer-rules.pro
@@ -1,0 +1,2 @@
+# Add specific rules for consumers of this library here.
+# This file will be included in the consumer's ProGuard configuration.
