+++ core/src/main/kotlin/com/example/core/CorePlaceholder.kt
@@ -1,0 +1,6 @@
+package com.example.core
+
+// This is a placeholder file to ensure the Kotlin directory structure is recognized.
+internal class CorePlaceholder {
+    fun getGreeting(): String = "Hello from CorePlaceholder"
+}
