+++ core/src/main/kotlin/com/example/core/data/local/db/UserEntity.kt
@@ -1,0 +1,12 @@
+package com.example.core.data.local.db // Updated package
+
+import androidx.room.Entity
+import androidx.room.PrimaryKey
+
+@Entity(tableName = "users")
+data class UserEntity(
+    @PrimaryKey(autoGenerate = true)
+    val id: Int = 0,
+    val name: String,
+    val email: String? = null
+)
