+++ core/src/main/kotlin/com/example/core/data/remote/ApiService.kt
@@ -1,0 +1,16 @@
+package com.example.core.data.remote
+
+// Import DTOs as needed, for example:
+// import com.example.core.data.remote.model.UserDto
+// import retrofit2.http.GET
+// import retrofit2.http.Path
+
+/**
+ * Interface for network API calls.
+ * Define suspend functions for each API endpoint.
+ */
+interface ApiService {
+    // Example endpoint:
+    // @GET("users/{userId}")
+    // suspend fun getUser(@Path("userId") userId: String): UserDto
+}
