+++ core/src/main/kotlin/com/example/core/di/DatabaseModule.kt
@@ -1,0 +1,28 @@
+package com.example.core.di
+
+import android.content.Context
+import com.example.core.data.local.db.AppDatabase // Correct import
+import com.example.core.data.local.db.UserDao // Correct import
+import dagger.Module
+import dagger.Provides
+import dagger.hilt.InstallIn
+import dagger.hilt.android.qualifiers.ApplicationContext
+import dagger.hilt.components.SingletonComponent
+import javax.inject.Singleton
+
+@Module
+@InstallIn(SingletonComponent::class)
+object DatabaseModule {
+
+    @Provides
+    @Singleton
+    fun provideAppDatabase(@ApplicationContext appContext: Context): AppDatabase {
+        return AppDatabase.getInstance(appContext)
+    }
+
+    @Provides
+    @Singleton // Daos are usually singletons tied to the database instance
+    fun provideUserDao(appDatabase: AppDatabase): UserDao {
+        return appDatabase.userDao()
+    }
+}
