+++ feature_welcome/src/main/kotlin/com/example/feature_welcome/Placeholder.kt
@@ -1,0 +1,5 @@
+package com.example.feature_welcome
+
+// This is a placeholder file to ensure the directory structure is created.
+// It can be removed once actual code is moved into this module.
+internal class FeatureWelcomePlaceholder
