+++ feature_welcome/src/main/kotlin/com/example/feature_welcome/WelcomePlaceholder.kt
@@ -1,0 +1,6 @@
+package com.example.feature_welcome
+
+// This is a placeholder file to ensure the Kotlin directory structure is recognized.
+internal class WelcomePlaceholder {
+    fun getGreeting(): String = "Hello from WelcomePlaceholder in feature_welcome"
+}
