+++ feature_welcome/src/main/kotlin/com/example/feature_welcome/domain/model/Message.kt
@@ -1,0 +1,5 @@
+package com.example.feature_welcome.domain.model
+
+data class Message(
+    val content: String
+)
