+++ feature_welcome/src/main/kotlin/com/example/feature_welcome/domain/repository/MessageRepository.kt
@@ -1,0 +1,9 @@
+package com.example.feature_welcome.domain.repository
+
+import com.example.feature_welcome.domain.model.Message // Updated import
+import kotlinx.coroutines.flow.Flow
+
+interface MessageRepository {
+    fun getWelcomeMessage(): Flow<Message>
+    suspend fun updateWelcomeMessage(newMessage: String)
+}
