+++ feature_welcome/src/main/kotlin/com/example/feature_welcome/domain/usecase/GetWelcomeMessageUseCase.kt
@@ -1,0 +1,12 @@
+package com.example.feature_welcome.domain.usecase // Updated package
+
+import com.example.feature_welcome.domain.model.Message // Updated import
+import com.example.feature_welcome.domain.repository.MessageRepository // Updated import
+import kotlinx.coroutines.flow.Flow
+import javax.inject.Inject
+
+class GetWelcomeMessageUseCase @Inject constructor(
+    private val messageRepository: MessageRepository
+) {
+    operator fun invoke(): Flow<Message> = messageRepository.getWelcomeMessage()
+}
